Analysis: Resolve sata extraction rules warning for sample app module
Warning Message: "The attribute android:allowBackup is deprecated from
Android 12 and higher and may be removed in future versions. Consider
adding the attribute android:dataExtractionRules specifying an @xml
resource which configures cloud backups and device transfers on
Android 12 and higher."

This 'android:allowBackup' attribute is not required for this
'SampleApp' app, it being an example app and all. Removing this
'android:allowBackup' and thus implicitly defaulting it to 'true'
wouldn't hurt.

For more info see:
- Back up user data with Auto Backup (Guide):
https://developer.android.com/guide/topics/data/autobackup