Build: Update dependency analysis configuration on source camera
This 'unused' related advise for the 'mediapicker.source-camera' module
is incorrect because otherwise Lint complains that
'androidx.core.content.FileProvider' cannot be found within the
'AndroidManifest.xml' file on that module. As such this advice is
instead suppressed and the 'onUnusedDependencies' exclusion
configuration were used to achieve that.

------------------------------------------------------------------------

Advice for :mediapicker:source-camera
Unused dependencies which should be removed:
  implementation 'androidx.core:core-ktx:1.12.0'