Build: Update dependency analysis configuration on source gif
This 'unused' related advise for the 'mediapicker.source-gif' module
is incorrect because otherwise the build fails with an
'NoClassDefFoundError' exception on 'GsonConverterFactory' (see below).
As such this advice is instead suppressed and the 'onUnusedDependencies'
exclusion configuration were used to achieve that.

------------------------------------------------------------------------

java.lang.NoClassDefFoundError: Failed resolution of:
 Lretrofit2/converter/gson/GsonConverterFactory;
 ...
 Caused by: java.lang.ClassNotFoundException: Didn't find class
  "retrofit2.converter.gson.GsonConverterFactory" on path:
  DexPathList[...]
 ...

------------------------------------------------------------------------

Advice for :mediapicker:source-gif
Unused dependencies which should be removed:
  implementation 'com.squareup.retrofit2:converter-gson:2.9.0'