Deps: Remove androidx core from sampleapp
This unused dependency got added to this project via this
(78fdbd38b12c6a4f2f6233135503db7a47de42c1) commit, which actually
created this module with an initial (default) setup. However, this
dependency never got to be used and as such can be safely removed.

------------------------------------------------------------------------

This removal was suggested by the dependency analysis report, see below:

Advice for :sampleapp
Unused dependencies which should be removed:
  implementation 'androidx.core:core-ktx:1.12.0'
  ...