Build: Update dependency analysis configuration on sampleapp
This 'unused' related advise for the 'sampleapp' module is actually
correct and 'mediapicker:source-wordpress' can be removed as a
dependency since it is actually unused and not being utilized from
within the sample app.

However, I chose to keep it there and suppress the advise instead in
order to preserve the sample app default configuration, its purpose
being to test all those individual modules:
- mediapicker:source-device
- mediapicker:source-gif
- mediapicker:source-wordpress
- mediapicker:source-camera

As such this advice is instead suppressed and the 'onUnusedDependencies'
exclusion configuration were used to achieve that.

------------------------------------------------------------------------

Advice for :sampleapp
Unused dependencies which should be removed:
  implementation project(':mediapicker:source-wordpress')